Refactored repositories and improved MemberGate.kt
