+++ src/main/kotlin/be/duncanc/discordmodbot/bot/commands/ReactionVote.kt
@@ -16,7 +16,7 @@
 
 package be.duncanc.discordmodbot.bot.commands
 
-import be.duncanc.discordmodbot.data.repositories.VotingEmotesRepository
+import be.duncanc.discordmodbot.data.repositories.jpa.VotingEmotesRepository
 import be.duncanc.discordmodbot.data.services.UserBlockService
 import net.dv8tion.jda.api.entities.Message
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
