+++ src/main/kotlin/be/duncanc/discordmodbot/bot/sequences/RevokeWarnPoints.kt
@@ -3,7 +3,7 @@ package be.duncanc.discordmodbot.bot.sequences
 import be.duncanc.discordmodbot.bot.commands.CommandModule
 import be.duncanc.discordmodbot.bot.services.MuteRole
 import be.duncanc.discordmodbot.data.entities.GuildWarnPoints
-import be.duncanc.discordmodbot.data.repositories.GuildWarnPointsRepository
+import be.duncanc.discordmodbot.data.repositories.jpa.GuildWarnPointsRepository
 import net.dv8tion.jda.api.MessageBuilder
 import net.dv8tion.jda.api.Permission
 import net.dv8tion.jda.api.entities.MessageChannel
