+++ src/main/kotlin/be/duncanc/discordmodbot/data/repositories/jpa/BlackListedWordRepository.kt
@@ -1,4 +1,4 @@
-package be.duncanc.discordmodbot.data.repositories
+package be.duncanc.discordmodbot.data.repositories.jpa
 
 import be.duncanc.discordmodbot.data.entities.BlackListedWord
 import be.duncanc.discordmodbot.data.entities.BlackListedWord.BlackListedWordId
