+++ src/main/kotlin/be/duncanc/discordmodbot/data/repositories/jpa/BlockedUserRepository.kt
@@ -14,15 +14,11 @@
  * limitations under the License.
  */
 
-package be.duncanc.discordmodbot.data.repositories
+package be.duncanc.discordmodbot.data.repositories.jpa
 
-import be.duncanc.discordmodbot.data.entities.MuteRole
+import be.duncanc.discordmodbot.data.entities.BlockedUser
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import org.springframework.transaction.annotation.Transactional
 
 @Repository
-interface MuteRolesRepository : JpaRepository<MuteRole, Long> {
-    @Transactional
-    fun deleteByRoleIdAndGuildId(roleId: Long, guildId: Long)
-}
+interface BlockedUserRepository : JpaRepository<BlockedUser, Long>
