+++ src/main/kotlin/be/duncanc/discordmodbot/data/repositories/jpa/ChannelOrderLockRepository.kt
@@ -1,4 +1,4 @@
-package be.duncanc.discordmodbot.data.repositories
+package be.duncanc.discordmodbot.data.repositories.jpa
 
 import be.duncanc.discordmodbot.data.entities.ChannelOrderLock
 import org.springframework.data.jpa.repository.JpaRepository
