+++ src/main/kotlin/be/duncanc/discordmodbot/data/repositories/jpa/VotingEmotesRepository.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package be.duncanc.discordmodbot.data.repositories
+package be.duncanc.discordmodbot.data.repositories.jpa
 
 import be.duncanc.discordmodbot.data.entities.VoteEmotes
 import org.springframework.data.jpa.repository.JpaRepository
