+++ src/main/kotlin/be/duncanc/discordmodbot/data/repositories/key/value/DiscordMessageRepository.kt
@@ -1,6 +1,6 @@
-package be.duncanc.discordmodbot.data.repositories
+package be.duncanc.discordmodbot.data.repositories.key.value
 
-import be.duncanc.discordmodbot.data.entities.DiscordMessage
+import be.duncanc.discordmodbot.data.redis.hash.DiscordMessage
 import org.springframework.data.keyvalue.repository.KeyValueRepository
 import org.springframework.stereotype.Repository
